chore: InlineCollection behavior is assigned to a separate SegmentsCollection type. This change is needed to give BlockCollection the ability to draw blocks.
